+++ assets/javascripts/kitten/components/molecules/info-lines/test.js
@@ -2,34 +2,33 @@ import React from 'react'
 import renderer from 'react-test-renderer'
 import { InfoLines } from './index'
 
-const createMockMediaMatcher = matches => () => ({
-  matches,
-  addListener: () => {},
-  removeListener: () => {},
-})
-
 describe('<InfoLines />', () => {
-  let originalMatchMedia
-
-  beforeEach(() => {
-    originalMatchMedia = window.matchMedia
-  })
-
-  afterEach(() => {
-    window.matchMedia = originalMatchMedia
-  })
-
   it('should match its empty snapshot', () => {
-    window.matchMedia = createMockMediaMatcher(false)
-
     const tree = renderer
       .create(
         <InfoLines
-          withBorderRadius={false}
-          withLeftRightBorder={false}
-          withoutTopBottomBorder={false}
-          withoutResponsive={false}
-          borderColor="#eee"
+          infos={[
+            { key: 'Lorem', value: 'Ipsum', id: 'item-1' },
+            { key: 'Dolor', value: 'Sit amet', id: 'item-2' },
+          ]}
+        />,
+      )
+      .toJSON()
+    expect(tree).toMatchSnapshot()
+  })
+
+  it('should match its propped snapshot', () => {
+    const tree = renderer
+      .create(
+        <InfoLines
+          withBorderRadius
+          withLeftRightBorder
+          withoutTopBottomBorder
+          withoutResponsive
+          borderColor="#f8f"
+          descriptionTagList="div"
+          titleTagList="span"
+          itemTagList="span"
           infos={[
             { key: 'Lorem', value: 'Ipsum', id: 'item-1' },
             { key: 'Dolor', value: 'Sit amet', id: 'item-2' },